Add a state test for the 09 precompile being defined
This test invokes delegatecall to the Blake2b (09) precompile in the
revision it has been defined (Istanbul) and in the one before it
(ConstantinopleFix). For the defined precompile the call should result
in a failure and the single gas unit provided should be consumed.
In case the precompile is undefined, the call should succeed and no gas
should be consumed.